Further setup of server
Wrapped things in a venv instance and set a requirements.txt
Got the first part of the Twitch API working correctly, it sends a
response to the api_redirect url with the code. But I'm currently stuck
on the next point, turning that into an access_token. Work to be
continued!